mood_cycle_supabase/requirements.txt
tushare>=1.2.89
pandas>=2.0.0
numpy>=1.24.0
flask>=2.3.0
flask-cors>=4.0.0
python-dateutil>=2.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
